backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.23
python-dotenv==1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0
requests>=2.31.0

